+++ audio_service/android/src/main/java/com/ryanheise/audioservice/AudioServicePlugin.java
@@ -88,13 +88,18 @@ public class AudioServicePlugin implements FlutterPlugin, ActivityAware {
                         }
                     }
                 }
-            } else if (context instanceof FlutterFragmentActivity) {
-                final FlutterFragmentActivity activity = (FlutterFragmentActivity)context;
-                Uri data = activity.getIntent().getData();
-                if (data != null) {
-                    initialRoute = data.getPath();
-                    if (data.getQuery() != null && !data.getQuery().isEmpty()) {
-                        initialRoute += "?" + data.getQuery();
+            } else if (context instanceof AudioServiceFragmentActivity) {
+                final AudioServiceFragmentActivity activity = (AudioServiceFragmentActivity)context;
+                initialRoute = activity.getInitialRoute();
+                if (initialRoute == null) {
+                    if (activity.shouldHandleDeeplinking()) {
+                        Uri data = activity.getIntent().getData();
+                        if (data != null) {
+                            initialRoute = data.getPath();
+                            if (data.getQuery() != null && !data.getQuery().isEmpty()) {
+                                initialRoute += "?" + data.getQuery();
+                            }
+                        }
                     }
                 }
             }